moving figures around
